requirements.txt
# Core dependencies
langchain==0.0.267
langgraph==0.0.17
openai==1.3.0
python-dotenv==1.0.0

# UI frameworks
streamlit==1.26.0
gradio==3.34.0

# Data handling
pandas>=2.0.3,<2.1.0
numpy>=1.24.3,<1.25.0
pydantic==2.1.1
sqlalchemy==2.0.20

# Utilities
python-dateutil==2.8.2
openpyxl==3.1.2